refactor: separate mapping logic from controller
